fix crash on deleting not yet added Log
In EasyLogsEditor, deleting a log during creation of it, when mLog is
null, caused the app to crash. A null check was introduced.

+++ app/src/main/java/com/example/sokol/monitor/EasyLogsDialog/EasyLogsEditor.java
@@ -126,6 +126,7 @@ public class EasyLogsEditor extends DialogFragment implements Dialog.OnClickList
     @Override
     public void onClick(DialogInterface dialogInterface, int i) {
         // deleting log
+        if(mLog == null) return;
         DbHelper db = DbHelper.getInstance(getActivity());
         db.deleteLog(mLog.getID());
         if(mListener == null) return;
